Introduce a base fragment for main login screens
