+++ WordPress/src/main/java/org/wordpress/android/modules/AppComponent.java
@@ -23,6 +23,7 @@ import org.wordpress.android.ui.accounts.SignInActivity;
 import org.wordpress.android.ui.accounts.SignInDialogFragment;
 import org.wordpress.android.ui.accounts.SignInFragment;
 import org.wordpress.android.ui.accounts.login.Login2FaFragment;
+import org.wordpress.android.ui.accounts.login.LoginBaseFormFragment;
 import org.wordpress.android.ui.accounts.login.LoginEmailFragment;
 import org.wordpress.android.ui.accounts.login.LoginEmailPasswordFragment;
 import org.wordpress.android.ui.accounts.login.LoginMagicLinkRequestFragment;
@@ -122,7 +123,7 @@ public interface AppComponent {
     void inject(NewUserFragment object);
 
     void inject(LoginActivity object);
-    void inject(LoginEmailFragment object);
+    void inject(LoginBaseFormFragment object);
     void inject(LoginMagicLinkRequestFragment object);
     void inject(LoginEmailPasswordFragment object);
     void inject(Login2FaFragment object);
