+++ WordPress/src/main/java/org/wordpress/android/ui/accounts/login/LoginBaseFormFragment.java
@@ -1,0 +1,187 @@
+package org.wordpress.android.ui.accounts.login;
+
+import android.app.ProgressDialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.support.annotation.LayoutRes;
+import android.support.annotation.Nullable;
+import android.support.annotation.StringRes;
+import android.support.v4.app.Fragment;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.text.TextWatcher;
+import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewStub;
+import android.widget.Button;
+import android.widget.TextView;
+
+import org.wordpress.android.R;
+import org.wordpress.android.WordPress;
+import org.wordpress.android.fluxc.Dispatcher;
+
+import javax.annotation.OverridingMethodsMustInvokeSuper;
+import javax.inject.Inject;
+
+public abstract class LoginBaseFormFragment extends Fragment implements TextWatcher {
+    private static final String KEY_IN_PROGRESS = "KEY_IN_PROGRESS";
+
+    private Button mPrimaryButton;
+    private Button mSecondaryButton;
+    private ProgressDialog mProgressDialog;
+
+    protected LoginListener mLoginListener;
+
+    private boolean mInProgress;
+
+    @Inject Dispatcher mDispatcher;
+
+    protected abstract @LayoutRes int getContentLayout();
+    protected abstract void setupLabel(TextView label);
+    protected abstract void setupContent(ViewGroup rootView);
+    protected abstract void setupBottomButtons(Button secondaryButton, Button primaryButton);
+    protected abstract @StringRes int getProgressBarText();
+
+    protected boolean isInProgress() {
+        return mInProgress;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        ((WordPress) getActivity().getApplication()).component().inject(this);
+
+        setHasOptionsMenu(true);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        ViewGroup rootView = (ViewGroup) inflater.inflate(R.layout.login_form_screen, container, false);
+        ViewStub form_container = ((ViewStub) rootView.findViewById(R.id.login_form_content_stub));
+        form_container.setLayoutResource(getContentLayout());
+        form_container.inflate();
+
+        setupLabel((TextView) rootView.findViewById(R.id.label));
+
+        setupContent(rootView);
+
+        mPrimaryButton = (Button) rootView.findViewById(R.id.primary_button);
+        mSecondaryButton = (Button) rootView.findViewById(R.id.secondary_button);
+        setupBottomButtons(mSecondaryButton, mPrimaryButton);
+
+        return rootView;
+    }
+
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        Toolbar toolbar = (Toolbar) view.findViewById(R.id.toolbar);
+        ((AppCompatActivity) getActivity()).setSupportActionBar(toolbar);
+
+        ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayShowTitleEnabled(false);
+            actionBar.setDisplayHomeAsUpEnabled(true);
+        }
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        if (savedInstanceState != null) {
+            mInProgress = savedInstanceState.getBoolean(KEY_IN_PROGRESS);
+
+            if (mInProgress) {
+                showProgressDialog();
+            }
+        }
+    }
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        if (context instanceof LoginListener) {
+            mLoginListener = (LoginListener) context;
+        } else {
+            throw new RuntimeException(context.toString() + " must implement LoginListener");
+        }
+    }
+
+    @Override
+    public void onDetach() {
+        super.onDetach();
+        mLoginListener = null;
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        mDispatcher.register(this);
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        mDispatcher.unregister(this);
+    }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+
+        outState.putBoolean(KEY_IN_PROGRESS, mInProgress);
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.menu_login, menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.help) {
+            mLoginListener.help();
+            return true;
+        }
+
+        return false;
+    }
+
+    protected void showProgressDialog() {
+        mPrimaryButton.setEnabled(false);
+        mSecondaryButton.setEnabled(false);
+
+        mProgressDialog =
+                ProgressDialog.show(getActivity(), "", getActivity().getString(getProgressBarText()), true, true,
+                        new DialogInterface.OnCancelListener() {
+                            @Override
+                            public void onCancel(DialogInterface dialogInterface) {
+                                if (mInProgress) {
+                                    endProgress();
+                                }
+                            }
+                        });
+        mInProgress = true;
+    }
+
+    @OverridingMethodsMustInvokeSuper
+    protected void endProgress() {
+        mInProgress = false;
+
+        if (mProgressDialog != null) {
+            mProgressDialog.cancel();
+            mProgressDialog = null;
+        }
+
+        mPrimaryButton.setEnabled(true);
+        mSecondaryButton.setEnabled(true);
+    }
+}
